moving csharp api to the api folder
