+++ journal/api/csharp/Program.cs
@@ -1,0 +1,39 @@
+
+namespace webapi {
+    public class Program {
+        public static void Main(string[] args) {
+            var builder = WebApplication.CreateBuilder(args);
+
+            // Add services to the container.
+
+            builder.Services.AddControllers();
+            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+            builder.Services.AddEndpointsApiExplorer();
+            builder.Services.AddSwaggerGen();
+
+            var app = builder.Build();
+
+            // Configure the HTTP request pipeline.
+            if (app.Environment.IsDevelopment()) {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+                app.UseDeveloperExceptionPage();
+            }
+            else {
+                app.UseExceptionHandler("/error");
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthorization();
+
+            app.MapGet("error", () => Results.Problem("An error occurred.")); //Minimal API for error handling
+
+            //app.MapGet("/error/test", () => { throw new Exception("test"); });
+
+            app.MapControllers();
+
+            app.Run();
+        }
+    }
+}
