+++ journal/api/csharp/appsettings.Development.json
@@ -1,0 +1,9 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Debug"
+    }
+  },
+  "UseDeveloperExceptionPage": true
+}
