+++ journal/api/csharp/webapi.http
@@ -1,0 +1,6 @@
+@webapi_HostAddress = http://localhost:5207
+
+GET {{webapi_HostAddress}}/People/GetAll
+Accept: application/json
+
+###
