--- journal/webapi/.mono/registry/CurrentUser/software/microsoft/csdevkit/telemetry/persistentpropertybag/microsoft.visualstudio.code.servicehost/values.xml
@@ -1,5 +1,0 @@
-﻿<values>
-  <value name="VS.TelemetryApi.DroppedEventsDuringDisposing" type="int">0</value>
-  <value name="VS.TelemetryApi.ChannelsDisposeLatency" type="int">1</value>
-  <value name="VS.TelemetryApi.TotalDisposeLatency" type="int">45</value>
-</values>
