--- journal/webapi/appsettings.Development.json
@@ -1,9 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Debug"
-    }
-  },
-  "UseDeveloperExceptionPage": true
-}
